made my chat box component on its own
